refactor: replace deprecated template provider usage
In Terraform v0.12 and above the `terraform_file` resource that is part
of the template provider has been replaced with a built-in function
`templatefile` that is simpler to use.  You can see this mentioned in
the docs at:
https://registry.terraform.io/providers/hashicorp/template/latest/docs/data-sources/file

You can also see that the `template` provider itself has been
archived: https://github.com/hashicorp/terraform-provider-template
There will therefore be no further updates or builds for it which
becomes a problem for Mac users on the new Apple Silicon chips, it
means that wherever the old and deprecated `template` provider is used
anyone on an Apple machine with M1 chip will not be able to
successfully run `terraform init`.

I'm not sure if there is a better way to iterate over all the
variables that need to be passed into the `templatefile` function as a
map.