feat: add iam role name optional variable
When using the latest version of the module the calculated default
name for the IAM role can get quite lengthy depending on the size of
your AWS account alias name.  A great number of our accounts end up
past the limit of 64 chars by the time the `org_name` and
`account_alias` are prepended.  Example error:

```
│ Error: expected length of name to be in the range (1 - 64), got matillion0-matillion-financial-commitments-read-bridgecrewcwssarole
│
│   with module.bridgecrew-read-only.aws_iam_role.bridgecrew_account_role,
│   on .terraform/modules/bridgecrew-read-only/iam_bridgecrew_role.tf line 2, in resource "aws_iam_role" "bridgecrew_account_role":
│    2:   name               = local.role_name
```

With this change it allows users to optionally set a custom IAM role
name to mitigate this issue or simply allow users to set what they
want inline with their usual role name conventions. If nothing is set
the naming should behave the same as it always has.